requirements.txt
Appium-Python-Client
selenium
adbutils
psutil
typing-extensions
aiofiles
tinydb